added data source and delegate for table view
